Split memento, playback, and contact out of view_common.py
